function_code/requirements.txt
# Azure Functions Core Tools
azure-functions==1.18.0

# Additional packages for your application
requests==2.31.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0